move modz to its own file and also move tets
